requirements.txt
# ── LangChain / LangGraph ──────────────────────────────────────────────────
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# ── OpenAI ─────────────────────────────────────────────────────────────────
openai>=1.50.0

# ── Vector DB ──────────────────────────────────────────────────────────────
pinecone[grpc]>=5.0.0          # rag/pinecone_client.py (PineconeGRPC)
# chromadb>=0.5.0              # Python 3.14 비호환 (Pydantic V1 충돌) — Pinecone 사용 시 불필요

# ── Web Search ─────────────────────────────────────────────────────────────
tavily-python>=0.5.0

# ── Backend ────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0      # config.py BaseSettings

# ── HTTP Client ────────────────────────────────────────────────────────────
httpx>=0.27.0                  # utils/llm.py 직접 호출

# ── PDF Processing ─────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Environment ────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Observability (Optional) ───────────────────────────────────────────────
langsmith>=0.1.0

# ── Testing ────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
